requirements.txt
torch
torchvision
timm
segmentation-models-pytorch
Pillow
opencv-python